+++ forgero-core/src/main/java/com/sigmundgranaas/forgero/core/property/v2/attribute/attributes/AttackSpeed.java
@@ -41,6 +41,6 @@ public class AttackSpeed implements Attribute {
 
     @Override
     public Float asFloat() {
-        return value;
+        return Math.max(value, -3.5f);
     }
 }