+++ forgero-fabric/src/main/java/com/sigmundgranaas/forgero/fabric/mixins/ItemStackAttributeDescriptionMixin.java
@@ -1,0 +1,9 @@
+package com.sigmundgranaas.forgero.fabric.mixins;
+
+import net.minecraft.item.ItemStack;
+import org.spongepowered.asm.mixin.Mixin;
+
+@Mixin(ItemStack.class)
+public class ItemStackAttributeDescriptionMixin {
+
+}