+++ minecraft-common/src/main/java/com/sigmundgranaas/forgero/minecraft/common/item/DefaultStateItem.java
@@ -6,8 +6,8 @@ import com.sigmundgranaas.forgero.core.state.StateProvider;
 import com.sigmundgranaas.forgero.core.type.Type;
 import com.sigmundgranaas.forgero.core.util.match.Context;
 import com.sigmundgranaas.forgero.core.util.match.Matchable;
-import com.sigmundgranaas.forgero.minecraft.common.item.tooltip.StateWriter;
-import com.sigmundgranaas.forgero.minecraft.common.item.tooltip.Writer;
+import com.sigmundgranaas.forgero.minecraft.common.tooltip.StateWriter;
+import com.sigmundgranaas.forgero.minecraft.common.tooltip.Writer;
 import net.minecraft.client.item.TooltipContext;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
@@ -36,7 +36,7 @@ public class DefaultStateItem extends Item implements StateItem, State {
 
     @Override
     public void appendTooltip(ItemStack itemStack, World world, List<Text> tooltip, TooltipContext tooltipContext) {
-        StateWriter.of(state(itemStack)).write(tooltip, tooltipContext);
+        StateWriter.of(dynamicState(itemStack)).write(tooltip, tooltipContext);
         super.appendTooltip(itemStack, world, tooltip, tooltipContext);
     }