+++ minecraft-common/src/main/java/com/sigmundgranaas/forgero/minecraft/common/tooltip/PassiveWriter.java
@@ -1,4 +1,4 @@
-package com.sigmundgranaas.forgero.minecraft.common.item.tooltip;
+package com.sigmundgranaas.forgero.minecraft.common.tooltip;
 
 import com.sigmundgranaas.forgero.core.property.passive.LeveledProperty;
 import com.sigmundgranaas.forgero.core.property.passive.PassiveProperty;